# backend/with-golang/api/internal/pkg/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample swagger for Fiber",
    "title": "Fiber Example API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "email": "fiber@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:3333",
  "basePath": "/",
  "paths": {
    "/v1/Example": {
      "get": {
        "consumes": ["application/json"],
        "tags": ["Example"],
        "summary": "Get all Example",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "page",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "userId",
            "name": "userId",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/entities.Example"
            }
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "tags": ["Example"],
        "summary": "Create Example",
        "parameters": [
          {
            "description": "query params",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/entities.Body"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/entities.Example"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "entities.Body": {
      "type": "object",
      "required": ["message"],
      "properties": {
        "message": {
          "type": "integer"
        }
      }
    },
    "entities.Example": {
      "type": "object",
      "required": ["id", "message"],
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "id": {
          "type": "integer"
        },
        "message": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "gorm.DeletedAt": {
      "type": "object",
      "properties": {
        "time": {
          "type": "string"
        },
        "valid": {
          "description": "Valid is true if Time is not NULL",
          "type": "boolean"
        }
      }
    }
  }
}